IoProvider: handle connect_error
When a socket fails to connect (for example, because the connection was
aborted from a middleware on the server), socket.io raises a
"connect_error" event, not an "error".

Handle this as well by reporting it through the error state variable.

See also: https://socket.io/docs/v4/middlewares/#handling-middleware-error